* Extract KryptonDesignerCollectionEditor to own file
Move KryptonDesignerCollectionEditor class from VisualDesignerCollectionForm.cs to its own dedicated file (KryptonDesignerCollectionEditor.cs) for better code organization and separation of concerns. The class functionality remains unchanged.
